@kortschak 2015-05-28
@btracey Please take a look. Particularly if there is any clarification in comments for the pool code.

I have added benchmarks for cases in master that are likely to be affected by this (blocked on #140 for Exp). I am not sure this is a win because my workstation is under heavy load and benchmarking is unreliable, so if you ran some that would be helpful - I will run some at home later.