@kortschak 2015-05-28
Benchmarks for the first commit.

```
name                             old mean              new mean              delta
MulDense100Half                   659µs × (0.99,1.01)   663µs × (0.98,1.01)    ~    (p=0.280 n=10+10)
MulDense100Tenth                  235µs × (0.97,1.03)   239µs × (0.97,1.04)    ~    (p=0.190 n=10+10)
MulDense1000Half                  586ms × (1.00,1.01)   588ms × (0.99,1.01)     ~    (p=0.095 n=9+10)
MulDense1000Tenth                 211ms × (0.99,1.01)   212ms × (0.99,1.00)      ~    (p=0.190 n=9+9)
MulDense1000Hundredth            86.7ms × (0.99,1.02)  86.2ms × (0.99,1.01)    ~    (p=0.190 n=10+10)
MulDense1000Thousandth           50.4ms × (0.99,1.01)  50.5ms × (0.99,1.01)     ~    (p=0.842 n=10+9)
PreMulDense100Half                610µs × (0.99,1.01)   612µs × (0.98,1.01)    ~    (p=0.315 n=10+10)
PreMulDense100Tenth               187µs × (0.96,1.03)   188µs × (0.96,1.04)     ~    (p=0.661 n=10+9)
PreMulDense1000Half               585ms × (1.00,1.00)   584ms × (1.00,1.01)     ~    (p=0.182 n=10+9)
PreMulDense1000Tenth              211ms × (0.99,1.01)   210ms × (1.00,1.00)     ~    (p=0.156 n=9+10)
PreMulDense1000Hundredth         84.7ms × (1.00,1.01)  84.8ms × (1.00,1.00)     ~    (p=0.356 n=10+9)
PreMulDense1000Thousandth        48.5ms × (0.99,1.01)  48.7ms × (0.99,1.01)    ~    (p=0.123 n=10+10)
Pow10_3                          7.72µs × (0.99,1.00)  7.08µs × (0.99,1.01)   -8.34% (p=0.000 n=10+9)
Pow100_3                         1.87ms × (1.00,1.00)  1.80ms × (1.00,1.01)  -3.39% (p=0.000 n=10+10)
Pow1000_3                         1.75s × (1.00,1.00)   1.76s × (0.99,1.01)  +0.80% (p=0.000 n=10+10)
Pow10_4                          10.9µs × (0.97,1.07)  10.1µs × (1.00,1.00)   -7.79% (p=0.000 n=10+9)
Pow100_4                         2.78ms × (0.99,1.02)  2.70ms × (1.00,1.00)   -2.96% (p=0.000 n=9+10)
Pow1000_4                         2.61s × (1.00,1.00)   2.63s × (1.00,1.00)   +0.78% (p=0.000 n=8+10)
Pow10_5                          10.6µs × (1.00,1.00)  10.1µs × (0.99,1.01)  -4.88% (p=0.000 n=10+10)
Pow100_5                         2.76ms × (1.00,1.00)  2.69ms × (0.99,1.01)   -2.41% (p=0.000 n=9+10)
Pow1000_5                         2.61s × (1.00,1.00)   2.63s × (1.00,1.00)  +0.71% (p=0.000 n=10+10)
Pow10_6                          13.7µs × (0.99,1.01)  13.1µs × (0.99,1.01)  -4.85% (p=0.000 n=10+10)
Pow100_6                         3.64ms × (1.00,1.00)  3.58ms × (0.99,1.01)  -1.54% (p=0.000 n=10+10)
Pow1000_6                         3.48s × (1.00,1.00)   3.50s × (1.00,1.00)   +0.69% (p=0.000 n=10+9)
Pow10_7                          13.6µs × (1.00,1.00)  13.0µs × (0.99,1.01)  -4.35% (p=0.000 n=10+10)
Pow100_7                         3.63ms × (1.00,1.00)  3.59ms × (0.99,1.01)   -1.12% (p=0.000 n=8+10)
Pow1000_7                         3.48s × (1.00,1.00)   3.51s × (0.99,1.00)  +0.70% (p=0.000 n=10+10)
Pow10_8                          16.6µs × (1.00,1.00)  16.0µs × (1.00,1.00)  -3.35% (p=0.000 n=10+10)
Pow100_8                         4.52ms × (1.00,1.00)  4.47ms × (0.99,1.01)  -1.11% (p=0.000 n=10+10)
Pow1000_8                         4.35s × (1.00,1.01)   4.39s × (1.00,1.01)   +0.96% (p=0.000 n=9+10)
Pow10_9                          24.2µs × (0.99,1.02)  15.0µs × (0.99,1.01)  -37.94% (p=0.000 n=19+9)
Pow100_9                         3.73ms × (0.99,1.01)  3.59ms × (0.99,1.01)  -3.74% (p=0.000 n=10+10)
Pow1000_9                         3.48s × (0.99,1.00)   3.52s × (0.99,1.01)   +1.06% (p=0.000 n=9+10)
MulWorkspaceDense100Half          959µs × (0.99,1.01)   904µs × (0.99,1.01)   -5.77% (p=0.000 n=9+10)
MulWorkspaceDense100Tenth         962µs × (0.99,1.02)   906µs × (0.99,1.01)   -5.86% (p=0.000 n=10+9)
MulWorkspaceDense1000Half         728ms × (1.00,1.00)   736ms × (0.99,1.01)    +1.03% (p=0.000 n=9+9)
MulWorkspaceDense1000Tenth        743ms × (1.00,1.00)   746ms × (1.00,1.00)  +0.47% (p=0.001 n=10+10)
MulWorkspaceDense1000Hundredth    790ms × (1.00,1.00)   793ms × (1.00,1.00)   +0.30% (p=0.004 n=9+10)
MulWorkspaceDense1000Thousandth  51.8ms × (0.94,1.05)  49.8ms × (0.95,1.06)  -3.90% (p=0.019 n=10+10)
```

I'm a little confused about old Pow10_9, but I ran extra runs of that and it's pretty solid.