@btracey 2015-05-28
What tool gives the statistical measures? If it's benchcmp the documentation hasn't been updated.