@btracey 2015-05-29
Here are the results from the pools. I must be doing something wrong with benchstat (I should probably look at some documentation) as it says everything has an n of 1+1 even though the tests were run for 10s

````
benchmark                                    old ns/op     new ns/op     delta      
BenchmarkMulWorkspaceDense100Half            684378        637115        -6.91%     
BenchmarkMulWorkspaceDense100Tenth           677266        616282        -9.00%     
BenchmarkMulWorkspaceDense1000Half           621878446     601740693     -3.24%     
BenchmarkMulWorkspaceDense1000Tenth          633495604     609224416     -3.83%     
BenchmarkMulWorkspaceDense1000Hundredth      625907040     613130711     -2.04%     
BenchmarkMulWorkspaceDense1000Thousandth     28396502      27508493      -3.13%     
````