@btracey 2015-05-29
I'm running the benchmarks properly, but the benchstat output makes no sense
````
brendan:~/Documents/mygo/src/github.com/gonum/matrix/mat64$ cat oldpool.txt 
PASS
BenchmarkMulWorkspaceDense100Half	   20000	    684378 ns/op
BenchmarkMulWorkspaceDense100Tenth	   20000	    677266 ns/op
BenchmarkMulWorkspaceDense1000Half	      30	 621878446 ns/op
BenchmarkMulWorkspaceDense1000Tenth	     100	 633495604 ns/op
BenchmarkMulWorkspaceDense1000Hundredth	     100	 625907040 ns/op
BenchmarkMulWorkspaceDense1000Thousandth	     500	  28396502 ns/op
ok  	github.com/gonum/matrix/mat64	204.817s
brendan:~/Documents/mygo/src/github.com/gonum/matrix/mat64$ cat newpool.txt 
PASS
BenchmarkMulWorkspaceDense100Half	   20000	    637115 ns/op
BenchmarkMulWorkspaceDense100Tenth	   20000	    616282 ns/op
BenchmarkMulWorkspaceDense1000Half	      30	 601740693 ns/op
BenchmarkMulWorkspaceDense1000Tenth	     100	 609224416 ns/op
BenchmarkMulWorkspaceDense1000Hundredth	     100	 613130711 ns/op
BenchmarkMulWorkspaceDense1000Thousandth	     500	  27508493 ns/op
ok  	github.com/gonum/matrix/mat64	196.600s
brendan:~/Documents/mygo/src/github.com/gonum/matrix/mat64$ benchstat oldpool.txt newpool.txt 
name                             old time/op  new time/op  delta
MulWorkspaceDense100Half          684µs ± 0%   637µs ± 0%   ~     (p=1.000 n=1+1)
MulWorkspaceDense100Tenth         677µs ± 0%   616µs ± 0%   ~     (p=1.000 n=1+1)
MulWorkspaceDense1000Half         622ms ± 0%   602ms ± 0%   ~     (p=1.000 n=1+1)
MulWorkspaceDense1000Tenth        633ms ± 0%   609ms ± 0%   ~     (p=1.000 n=1+1)
MulWorkspaceDense1000Hundredth    626ms ± 0%   613ms ± 0%   ~     (p=1.000 n=1+1)
MulWorkspaceDense1000Thousandth  28.4ms ± 0%  27.5ms ± 0%   ~     (p=1.000 n=1+1)
```` 
The benchmark files look fine, and benchstat agrees that there is a very small error on the numbers. And yet, is reports a statistical insignificance?