@btracey 2015-05-29
Here are the noasm numbers 
````
brendan:~/Documents/mygo/src/github.com/gonum/matrix/mat64$ benchstat oldpoolnoasm.txt newpoolnoasm.txt 
name                             old time/op  new time/op  delta
MulWorkspaceDense100Half         1.47ms ± 5%  1.51ms ± 5%     ~     (p=0.240 n=6+6)
MulWorkspaceDense100Tenth        1.51ms ± 2%  1.69ms ±16%  +11.68%  (p=0.026 n=6+6)
MulWorkspaceDense1000Half         1.08s ± 6%   1.15s ± 1%   +6.49%  (p=0.009 n=6+5)
MulWorkspaceDense1000Tenth        1.25s ± 5%   1.22s ± 2%   -2.68%  (p=0.015 n=6+6)
MulWorkspaceDense1000Hundredth    1.24s ± 0%   1.33s ± 2%   +7.17%  (p=0.010 n=4+6)
MulWorkspaceDense1000Thousandth  31.5ms ±23%  37.8ms ±39%     ~     (p=0.093 n=6+6)
````

It's probably just noise from my computer doing other things. 

I don't think this is a real test of whether the asm code in here does anything or not. The time for Mul itself is highly affected by asm because of the assembly used in the inner loop of the Mul operation.